# contracts/openapi/order-api-v1.yaml
openapi: 3.1.0
info:
  title: Polaris Order API
  version: 1.0.0
  description: Public order API exposed through the Polaris gateway.
servers:
  - url: http://localhost:8080
    description: Local gateway
security:
  - bearerAuth: []
tags:
  - name: Orders
paths:
  /api/v1/orders:
    post:
      tags:
        - Orders
      operationId: placeOrder
      summary: Place an order
      description: >-
        The customer ID is taken from the UUID JWT subject. The token requires the `orders:write` scope.
        A pending order is committed before inventory reservation. A failed response may leave an order
        that background recovery later completes; reuse Idempotency-Key when retrying to avoid another order.
      x-required-scope: orders:write
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: >-
            Replays the original result for the same customer and payload, including after a failed response.
            Reuse with a different payload returns 409. Without this header each HTTP attempt creates a distinct order.
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order created and synchronously confirmed or cancelled according to inventory.
          headers:
            Location:
              description: Relative URI of the created order.
              schema:
                type: string
                format: uri-reference
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Idempotency-Replayed:
              description: Whether this response replays a completed request with the same idempotency key.
              required: true
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: A valid bearer token is required by the gateway.
        '403':
          description: The bearer token does not have the `orders:write` scope.
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/InventoryUnavailable'
  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      operationId: getOrder
      summary: Get an order
      description: Returns an order owned by the UUID JWT subject. The token requires the `orders:read` scope.
      x-required-scope: orders:read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order found.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: A valid bearer token is required by the gateway.
        '403':
          description: The bearer token does not have the `orders:read` scope.
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestId:
      description: Correlation ID accepted or generated by the gateway and order service.
      schema:
        type: string
  schemas:
    PlaceOrderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemRequest:
      type: object
      required:
        - sku
        - quantity
        - unitPrice
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 128
        quantity:
          type: integer
          format: int32
          minimum: 1
        unitPrice:
          type: number
          minimum: 0.01
    Order:
      type: object
      additionalProperties: false
      required:
        - id
        - customerId
        - status
        - items
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      additionalProperties: false
      required:
        - id
        - sku
        - quantity
        - unitPrice
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
    OrderStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - CANCELLED
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
    OrderNotFoundProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required:
            - orderId
          properties:
            orderId:
              type: string
              format: uuid
    RateLimitError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          const: rate_limit_exceeded
  responses:
    BadRequest:
      description: The request body, path parameter, JSON representation, or idempotency key is invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    OrderNotFound:
      description: No order exists for the supplied ID.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/OrderNotFoundProblem'
    InventoryUnavailable:
      description: >-
        Inventory reservation could not be resolved during the request. The persisted pending order may
        complete through background recovery. Retry with the same Idempotency-Key and payload to recover its result.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    IdempotencyConflict:
      description: The idempotency key was already used by this customer for a different request payload.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: The gateway rate limit was exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'